clean up code, split to spectrum.py and performance.py
